requirements.txt
virustotal-api
inotify
requests
elasticsearch
